requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
elasticsearch==8.17.2
elasticsearch-dsl==8.17.1
python-dotenv==1.0.1